Write symbol table

+++ dependencies/mini-rgx/src/mini-rgx.h
@@ -1,12 +1,12 @@
+#include <string>
 #include <optional>
 
 namespace mrx {
-
     struct substr {
-        unsigned long a;
-        unsigned long b;
+        size_t a;
+        size_t b;
 
-        inline int end() const {
+        inline size_t end() const {
             return a + b;
         }
     };
@@ -24,6 +24,6 @@ namespace mrx {
     //Find a certain sequence of characters with a regex, returns the first matching substring as a string.
     extern std::optional<std::string> FindRgxSubstr(const std::string string, const char* pattern, const unsigned short start = 0);
 
-    //FindRgx, with sections or "tokens". See the documentation for more.
+    //FindRgx, with sections/tokens. See the documentation for more.
     extern unsigned char FindRgxSectional(substr(&out_sections)[], const std::string str, const char* pattern, const char(&invalid)[] = "");
 }
